refactor: String extension MainViewController로 이동하고 fileprivate 설정
