+++ Expo1900/Expo1900/Controller/MainViewController.swift
@@ -71,3 +71,29 @@ final class MainViewController: UIViewController {
     }
 }
 
+fileprivate extension String {
+    var expoVisitorInformation: String {
+        return ExpoConstant.visitor + ExpoConstant.colon + self + ExpoConstant.people
+    }
+    
+    var expoLocationInformation: String {
+        return ExpoConstant.location + ExpoConstant.colon + self
+    }
+    
+    var expoDurationInformation: String {
+        return ExpoConstant.duration + ExpoConstant.colon + self
+    }
+    
+    func createAttributed(target: String) -> NSAttributedString {
+        let fontSize = ExpoConstant.mediumFont
+        let attributed: NSMutableAttributedString = NSMutableAttributedString(string: self)
+        let targetRange = (self as NSString).range(of: target)
+        attributed.addAttribute(.font, value: fontSize, range: targetRange)
+        
+        return attributed
+    }
+    
+    func applyLineBreak() -> String {
+        return self.split(separator: ExpoConstant.leftBracket).joined(separator: "\n(")
+    }
+}
